+++ demo/scripts/rxPermalink.js
@@ -4,11 +4,12 @@ angular.module('demoApp')
         restrict: 'E',
         scope: {},
         transclude: true,
-        template: '<a ng-href="{{permalinkUrl}}" class="permalink" id="{{permaId}}" ng-transclude></a>',
+        templateUrl: 'templates/rx-permalink.html',
         link: function (scope, el, attr) {
             // Convert Text content to URL-safe string
             scope.permaId = el.text()
                 .toLowerCase() // normalize to lowercase
+                .trim() // remove leading/trailing whitespace
                 .replace(/[^0-9a-z]/gi, '-') // non-alphanumeric to hyphen
                 .replace(/\-{2,}/g, '-'); // condense consecutive hyphens