+++ src/utilities/rxAutoSave/scripts/rxAutoSave.spec.js
@@ -27,7 +27,7 @@ describe('utilities:rxAutoSave', function () {
     };
 
     beforeEach(function () {
-        module('encore.ui.utilities');
+        module('encore.ui.elements');
 
         module(function ($provide) {
             $provide.value('$location', $location);